// src/lib2/models/PlanComptableDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PlanComptableDto = {
    id?: string;
    libelle: string;
    classe?: number;
    notes?: string;
    actif?: boolean;
    noCompte: string;
    createdAt?: string;
    updatedAt?: string;
    createdBy?: string;
    updatedBy?: string;
};